requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argcomplete==3.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.0
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
clarabel==0.10.0
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cvxopt==1.3.2
cvxpy==1.6.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
diskcache==5.6.3
dlai_tools==0.6.1
dnspython==2.6.1
executing==2.2.0
fancyimpute==0.7.0
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.0.3
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.50.0
fpdf==1.7.2
fqdn==1.5.1
fsspec==2025.3.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.6
imbalanced-learn==0.13.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.9.2
kiwisolver==1.4.5
kneed==0.8.5
knnimpute==0.1.0
langchain==0.3.26
langchain-core==0.3.67
langchain-text-splitters==0.3.8
langsmith==0.4.4
libclang==18.1.1
llama_cpp_python==0.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
ml_dtypes==0.5.1
mongoengine==0.28.2
mpmath==1.3.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nose==1.3.7
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
ollama==0.5.1
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.18
osqp==0.6.7.post3
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pgmpy==0.1.26
pillow==10.3.0
pipx==1.4.3
platformdirs==4.2.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.5.2
Pygments==2.19.1
pymongo==4.6.2
pyparsing==3.1.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
qdldl==0.1.7.post5
referencing==0.36.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
scs==3.2.7.post2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==69.2.0
six==1.16.0
sklearn-compat==0.1.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
userpath==1.9.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.1
wheel==0.45.1
wrapt==1.17.2
xgboost==2.1.4
zstandard==0.23.0
